Improve type corectness for text functions

+++ inc/utils.h
@@ -77,7 +77,7 @@ static inline char *log_buffer(const message_log *log)
 }
 
 /* text array handling */
-GPtrArray *text_wrap(const char *str, int width, int indent);
+GPtrArray *text_wrap(const char *str, unsigned width, unsigned indent);
 
 /**
  * append one array of text to another array of text.
@@ -94,7 +94,7 @@ GPtrArray *text_append(GPtrArray *first, GPtrArray *second);
  * @param text An array of strings
  * @return the length of the longest string in the array
  */
-int text_get_longest_line(GPtrArray *text);
+size_t text_get_longest_line(GPtrArray *text);
 
 void text_destroy(GPtrArray *text);
